RunSwarm simplified
Duplicative code removed from RunSwarm dealing with accurately labelling OTU representative sequences when curated ESV sequences are the input. Specifically checking that ESVs are marked as both ESVrepresentative sequences and OTUrepresentativesequences in order to be accepted as OTUrepresentative sequences is not necessary as this must be the case based on how these assignments are generated.