Idtaxa outputs NAs at unclassified ranks
IDtaxa's default behaviour is to cname the next lowest rank unclassified x and then to insert NAs at ranks below that. The pipeline now outputs in keeping with this standard.

This behaviour is preferable as seqs unassigned by Idtaxa at a rank (either due to the reference seq being unassigned at that rank or there being no reference seq which can be assigned to at the chosen confidence threshold)  are now clearly labelled NA. And if a higher ranks classification is desired this can be easliy called by for example asking for the class assignment of all sequences unassigned at genus.

Note previous indecisiveness on this issue with multiple previous commits exploring this behaviour