DADA2: chimera removal moved after run merging
the tracking table needed to be modified to account for this change. non-chimeric read counts are now taken once all runs are merged, as a result were samples are run multiple times and merged there is not a non-chimeric read count for each sample run but only for the merged sample overall. NAs are entered in duplicate samples and the table is sorted by sample name to make clear where merging has occurred.